backend/requirements.txt
# Requirements for Backend - Place in backend/requirements.txt

# Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# Async
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.3.0

# Optional: For production
gunicorn==21.2.0
slowapi==0.1.9  # Rate limiting

# Optional: For advanced features
celery==5.3.4  # Background tasks
redis==5.0.1  # Caching
pandas==2.1.3  # Data processing
openpyxl==3.1.1  # Excel support
python-dateutil==2.8.2  # Date utilities
